requirements.txt
black
Faker~=37.1.0
scikit-learn~=1.6.1
nltk~=3.9.1
gender-guesser~=0.4.0
fastapi~=0.115.12
uvicorn~=0.34.0
streamlit~=1.44.1
requests~=2.32.3
sentence-transformers~=4.0.2
spacy~=3.8.5
pydantic~=2.11.3
pdfplumber~=0.11.6
python-docx~=1.1.2